feat(Interaction): add new door controllable replacing 3d control
The 3D Control Door has now been deprecated for the new Physics Door
Controllable. It offers a simpler solution that enables the script to
be enabled and disabled at runtime with correct clean up.

It no longer auto detects direction of door opening and requires a
specific placement of a Transform object when there hinge is to be
created. This removes any glitches caused by calculations and also
allows finer placement of the hinge.

Due to the way the new Physics Door script works means the Chest
can also be easily recreated with the Physics Door script, this means
that the previous 3D Control Chest script has also been deprecated in
favour of building chests using this new Physics Door script.